st/oxr: Add OXR_TWO_CALL_HELPER_DIRECT macro
This macro can be used if the requested information is not available
in a local array and has to be dynamically built.
In this case it is more convenient to write directly into the output
array instead of allocating a temporary array.
However the caller has to handle the capacity correctly.